+++ CHANGELOG.md
@@ -5,7 +5,9 @@
 - Fix compilation errors when included as lib in a new Angular 13.3.0 project [mpellerin42]
 - Fix side nav item header's color [mpellerin42]
 - Fix background color for disabled inverted solid buttons [mpellerin42]
-- Fix chips demo [mpellerin42]
+- Fix chips: [mpellerin42]
+  - expandable chips: don't trigger expanded event on disabled chip
+  - demo: fix the description and add documentation on expandable chips 
 
 ### Improvements
 - **Typography**:  [mpellerin42]
@@ -19,7 +21,7 @@
 - **Scrollbars**: Add tokens for scrollbar border radius and distinction between active and inactive state [mpellerin42]
 - **Chips**: [mpellerin42]
   - Tokenize chip colors and borders
-  - 
+  - Add `backgroundColor`, `textColor` and `borderColor` inputs to all the chips
 
 # 2.50.4 (2022-06-13)